# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Accessibility API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
...
      "Organization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "limits": {
            "$ref": "#/components/schemas/PlanLimits"
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "plan",
          "limits"
        ]
      },
      "MembershipRole": {
        "type": "string",
        "enum": [
          "owner",
          "member"
        ]
      },
      "Plan": {
        "type": "string",
        "enum": [
          "free",
          "pro"
        ]
      },
      "PlanLimits": {
        "type": "object",
        "properties": {
          "maxSites": {
            "type": "integer",
            "minimum": 1
          },
          "maxPagesPerAudit": {
            "type": "integer",
            "minimum": 1
          },
          "scheduledFrequencies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScanFrequency"
            }
          },
          "manualAuditsPerDayPerSite": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "maxSites",
          "maxPagesPerAudit",
          "scheduledFrequencies",
          "manualAuditsPerDayPerSite"
        ]
      },
      "ScanFrequency": {
        "type": "string",
        "enum": [
          "weekly",
          "daily"
        ]
      },
...
      "CreateOrganizationRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "name"
        ]
      },
      "Site": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orgId": {
            "type": "string",
            "format": "uuid"
          },
          "baseUrl": {
            "type": "string",
            "format": "uri"
          },
          "verifiedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "verificationMethod": {
            "$ref": "#/components/schemas/VerificationMethod"
          },
          "scanFrequency": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "weekly",
              "daily",
              null
            ]
          },
          "nextScanAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "verification": {
            "type": "object",
            "properties": {
              "dnsRecord": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "TXT"
                    ]
                  },
                  "name": {
                    "type": "string"
                  },
                  "value": {
                    "type": "string"
                  }
                },
                "required": [
                  "type",
                  "name",
                  "value"
                ]
              },
              "file": {
                "type": "object",
                "properties": {
                  "path": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                },
                "required": [
                  "path",
                  "content"
                ]
              }
            },
            "required": [
              "dnsRecord",
              "file"
            ]
          }
        },
        "required": [
          "id",
          "orgId",
          "baseUrl",
          "verifiedAt",
          "verificationMethod",
          "scanFrequency",
          "nextScanAt",
          "verification"
        ]
      },
      "VerificationMethod": {
        "type": [
          "string",
          "null"
        ],
        "enum": [
          "dns",
          "file",
          null
        ]
      },
      "CreateSiteRequest": {
        "type": "object",
        "properties": {
          "baseUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "baseUrl"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/orgs": {
      "post": {
        "summary": "Create an organization owned by the caller",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Organization"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Organizations the caller belongs to",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Organizations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Organization"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/{orgId}/sites": {
      "post": {
        "summary": "Add a site to an organization",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSiteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Site"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The organization's plan allows no more sites",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Site already added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Sites of an organization",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Sites",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Site"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}